requirements.txt
Flask
line-bot-sdk
pymongo[srv]
python-dotenv
firebase-admin
openai